+++ android/shadps4-app/core/runtime/src/main/kotlin/com/shadps4/android/runtime/session/NativeFexSession.kt
@@ -21,7 +21,7 @@ object NativeFexSession {
     external fun nativeSaveDialogRespond(generation: Long, request: Long, action: Int, selection: Int): Boolean
 
     external fun nativeSetGuestShadingQuality(quality: Int)
-    external fun nativeSetInternalScalePercent(percent: Int)
+    external fun nativeSetInternalScalePercent(percent: Float)
     external fun nativeSetConsoleLanguage(language: Int)
 
     external fun nativeIdentity(): String
